commit task

+++ sort.h
@@ -31,7 +31,7 @@ void bubble_sort(int *array, size_t size);
 void insertion_sort_list(listint_t **list);
 void selection_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
-void quicksort_rec(int *array, int low, int high, size_t size);
-int lomuto_par(int *array, int low, int high, size_t size);
+void quicksort_rec(int *array, size_t low, size_t high, size_t size);
+size_t lomuto_par(int *array, size_t low, size_t high, size_t size);
 
 #endif
